+++ app/components/CommentSection/styles.ts
@@ -52,6 +52,18 @@ const styles = StyleSheet.create({
     color: '#666',
     marginBottom: 16,
   },
+  input: {
+    height: 40,
+    borderColor: 'gray',
+    borderWidth: 1,
+    marginBottom: 12,
+    paddingHorizontal: 8,
+  },
+  buttonContainer: {
+    marginBottom: 20,
+    borderRadius: 10,
+    overflow: 'hidden',
+  },
 })
 
 export default styles
