--- app/components/CreateComment/styles.ts
@@ -1,21 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-const styles = StyleSheet.create({
-  container: {
-    padding: 16,
-  },
-  input: {
-    height: 40,
-    borderColor: 'gray',
-    borderWidth: 1,
-    marginBottom: 12,
-    paddingHorizontal: 8,
-  },
-  buttonContainer: {
-    marginBottom: 20,
-    borderRadius: 10,
-    overflow: 'hidden',
-  },
-})
-
-export default styles
